+++ src/App.tsx
@@ -1,14 +1,13 @@
-// Renders the stopwatch and handles its functionality
-
 import React, { useEffect, useState } from 'react';
 import DisplayComponent from './StopWatch';
 import { LapButtonComponent, ResetButtonComponent, StartButtonComponent, StopButtonComponent } from './StopWatchButton';
 import './App.css';
 
+// Renders the stopwatch and handles its functionality
 export default function App() {
     // State hooks for tracking time, counting status, and lap times
-    const [time, setTime] = useState(0); 
-    const [counting, setCounting] = useState(false); 
+    const [time, setTime] = useState<number>(0); 
+    const [counting, setCounting] = useState<boolean>(false); 
     const [laps, setLaps] = useState<number[]>([]);
 
     // Effect hook for handling the stopwatch functionality
@@ -44,13 +43,19 @@ export default function App() {
 
             {/* Displays button components*/}
             <div className="buttons">
-                {/* Shows the Start and Reset button when the stopwatch is not running */}
-                {!counting && <StartButtonComponent setCounting={setCounting} />}
-                {!counting && <ResetButtonComponent setTime={setTime} setLaps={setLaps} />}
-                
-                {/* Shows the Stop and Lap buttons when the stopwatch is running */}
-                {counting && <StopButtonComponent setCounting={setCounting} />}
-                {counting && <LapButtonComponent onLap={handleLap} />}
+                {counting ? (
+                    // Show Stop and Lap buttons when the stopwatch is running
+                    <>
+                        <StopButtonComponent setCounting={setCounting} />
+                        <LapButtonComponent onLap={handleLap} />
+                    </>
+                ) : (
+                    // Show Start and Reset buttons when the stopwatch is not running
+                    <>
+                        <StartButtonComponent setCounting={setCounting} />
+                        <ResetButtonComponent setTime={setTime} setLaps={setLaps} />
+                    </>
+                )}
             </div>
 
             {/* Displays recorded laps */}
